+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -321,7 +321,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT7H3M8MBT000DYA46NQDHSD"
+        "01KT8T9AEVEH1WTBX1RTCBD7Z3"
     )
     r"""The unique identifier of the data source"""