+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT7H3M8HRB540MQJFDWS8ZM5"
+        "01KT8T9AERH8SB44EK57G7VHCF"
     )
     r"""The unique identifier of the data source"""