+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT7H3M8KKH9B8P1MDA55XDZJ"
+        "01KT8T9AETWT4CHJGPJ2C5046A"
     )
     r"""The unique identifier of the data source"""