+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT7H3M8Q9450VA0WQW3EW33B"
+        "01KT8T9AEYJXQEQHRYSXVR4KAQ"
     )
     r"""The unique identifier of the data source"""