+++ packages/orq-rc/RELEASES.md
@@ -8449,3 +8449,13 @@ Based on:
 - [python v4.10.0-rc.82] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.82] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.82 - packages/orq-rc
+
+## 2026-06-04 08:01:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.770.0 (2.893.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.83] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.83] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.83 - packages/orq-rc